feat(neon): Add feature flag for enabling macros
This allows making the `linkme` dependency optional which may be problematic on some platforms.
