+++ crates/neon-macros/Cargo.toml
@@ -14,3 +14,7 @@ proc-macro = true
 proc-macro2 = "1.0.79"
 quote = "1.0.33"
 syn = { version = "2.0.57", features = ["full"] }
+
+[features]
+# Enable `#[neon::export]`
+export = []
